removed chatlist from emergency tabs
