+++ Screens/First-Aid Screens/Hypothermia.js
@@ -4,12 +4,14 @@ import { Container, Text, Content, Card, CardItem, Body, Left, View } from "nati
 import YoutubePlayer from 'react-native-youtube-iframe';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import { StyleSheet } from "react-native";
-import Data from "../Data/Burns.json";
-export default function Burns() {
+import Data from "../../Data/Hypothermia.json"
+export default function Hypothermia() {
 
   const navigation = useNavigation();
-
-  var res = Data.filter(function (item) {
+    function HypothermiaNav() {
+    navigation.navigate("Home");
+  }
+  var res = Data.filter(function(item) {
     return item.id;
   });
   return (
@@ -53,23 +55,12 @@ export default function Burns() {
                   </View>)
               })}
             </View>
-            <View style={styles.content}>
-              <MaterialCommunityIcons name="numeric-4-circle" size={25} />
-              {Data.map(data => {
-                return (
-                  <View key={data.id}>
-                    <Text>
-                      {res[3].text}
-                    </Text>
-                  </View>)
-              })}
-            </View>
           </CardItem>
-          <View >
+          <View>
             <YoutubePlayer
               height={300}
               play={false}
-              videoId={'DzpRjE5ekVk'}
+              videoId={'dEKaCOx7igI'}
             />
           </View>
         </Card>
@@ -77,6 +68,7 @@ export default function Burns() {
     </Container>
   );
 }
+
 const styles = StyleSheet.create({
   content: {
     marginRight: 4,
@@ -95,3 +87,4 @@ const styles = StyleSheet.create({
     justifyContent: 'center'
   }
 });
+
