+++ Screens/First-Aid Screens/Poisoning.js
@@ -4,12 +4,12 @@ import { Container, Text, Content, Card, CardItem, Body, Left, View } from "nati
 import YoutubePlayer from 'react-native-youtube-iframe';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import { StyleSheet } from "react-native";
-import Data from "../Data/Meningitis.json"
-export default function Meningitis() {
+import Data from "../../Data/Poisoning.json"
+export default function Poisoning() {
 
   const navigation = useNavigation();
 
-  function MeningitisNav() {
+  function PoisoningNav() {
     navigation.navigate("Home");
   }
   var res = Data.filter(function(item) {
@@ -27,7 +27,7 @@ export default function Meningitis() {
               <MaterialCommunityIcons name="numeric-1-circle" size={25} />
               {Data.map(data => {
                 return (
-                  <View key={data.id}>
+                  <View style={{flex=1}} key={data.id}>
                     <Text>
                       {res[0].text}
                     </Text>
@@ -45,12 +45,23 @@ export default function Meningitis() {
                   </View>)
               })}
             </View>
+            <View style={styles.content}>
+              <MaterialCommunityIcons name="numeric-3-circle" size={25} />
+              {Data.map(data => {
+                return (
+                  <View key={data.id}>
+                    <Text>
+                      {res[2].text}
+                    </Text>
+                  </View>)
+              })}
+            </View>
           </CardItem>
           <View>
             <YoutubePlayer
               height={300}
               play={false}
-              videoId={'su_MyX6BG6A'}
+              videoId={'b2ieb8BZJuY'}
             />
           </View>
         </Card>
@@ -58,23 +69,6 @@ export default function Meningitis() {
     </Container>
   );
 }
-
-const card = {
-  width: 350,
-  height: 150
-}
-const text = {
-  textAlign: 'center',
-  fontWeight: 'bold',
-  color: '#5d1a0c',
-  fontSize: 18,
-}
-const button = {
-  width: 50,
-  height: 50,
-  alignContent: 'center',
-  justifyContent: 'center'
-};
 
 const styles = StyleSheet.create({
   content: {
@@ -94,4 +88,3 @@ const styles = StyleSheet.create({
     justifyContent: 'center'
   }
 });
-
